Using mouseClicked on mobile. Basic functionality should now be working on mobile, although I'd like to figure out how to make an exception for iOS and iPad since it wasn't having the issues android was.
Group creation on mobile is now done by clicking a starting character, then clicking an ending character.
Everything else is the same.

Also hid the saveScramble button when the default text is being shown. Stops the user from saving an overly long message that was too wide to be deleted!